minor refactor with small changes
the only thing i changed is that we shouldn't store the actual config in
the appdata files, we should store references to them, the same is true
for github repos.

From the refactoring side, i moved most of the logic into the appdata
crate, because for example loading from the repo or from a local file
is required for a lot of commands!

The appdata is now encoded in the AppDataData and works over serde
serialize, which makes it a lot easier to edit.

i also added a file and branch option to the select command, where a
branch and a file inside of the repository can be selected!

i also moved the pulling logic from the "config select <github-url>"
command into a seperate command "config pull" because sometimes we might
not have internet. Especially at EHW internet access is not allowed
during demonstrations.